BRANCH BASE-370a:

+++ src/tactic/command/watch_drop_folder.py
@@ -20,7 +20,6 @@ import time, os, shutil, sys
 import os.path
 import sys
 
-from dateutil import parser
 from pyasm.common import Environment, Config, Common
 from pyasm.security import Batch
 from pyasm.biz import Project
@@ -31,12 +30,13 @@ from tactic.command import SchedulerTask, Scheduler
 from time import gmtime, strftime
 from optparse import OptionParser
 from tactic.command import PythonCmd
+from dateutil import parser
 
 import threading
 
 
 import logging
-logging.basicConfig(filename='/tmp/myapp.log', level=logging.INFO)
+#logging.basicConfig(filename='/tmp/myapp.log', level=logging.INFO)
 
 try:
     from watchdog.observers import Observer
@@ -76,46 +76,60 @@ class TestLoggingEventHandler(LoggingEventHandler):
 
 
 class WatchFolderFileActionThread(threading.Thread):
-
+    
     def __init__(my, **kwargs):
         my.kwargs = kwargs
         super(WatchFolderFileActionThread, my).__init__()
 
     def run(my):
 
-        Batch()
+        task = my.kwargs.get("task")
+        
+        project_code = task.project_code
+        Batch(project_code=project_code)
         try:
             my._run()
         finally:
             task = my.kwargs.get("task")
             paths = task.get_paths()
+            task.set_clean(True)
             for path in paths:
-                checkin_path = "%s.lock" % path
-                if os.path.exists(checkin_path):
-                    os.unlink(checkin_path)
+                lock_path = "%s.lock" % path
+                if os.path.exists(lock_path):
+                    os.unlink(lock_path)
+            task.set_clean(False)
 
 
     def _run(my):
 
         task = my.kwargs.get("task")
         paths = task.get_paths()
-        
         count = 0
         restart = False
 
         while True:
+            
             if not paths:
                 time.sleep(1)
                 continue
-
+            
             path = paths.pop(0)
-
             checkin_path = "%s.checkin" % path
+            lock_path = "%s.lock" % path
             error_path = "%s.error" % path
 
-            if not os.path.exists(checkin_path):
-                print "ERROR: no lock path [%s]" % checkin_path
+            if not os.path.exists(path):
                 continue
+            if not os.path.exists(checkin_path):
+                #print "Action Thread SKIP: no checkin path [%s]" % checkin_path
+                continue
+            else:
+                # Exit if another process is also checking this file in.                
+                f = open(checkin_path, "r")
+                pid = f.readline()
+                f.close()
+                if pid != str(os.getpid()):
+                    continue
 
             try:
 
@@ -124,15 +138,18 @@ class WatchFolderFileActionThread(threading.Thread):
                     "search_type": task.search_type,
                     "base_dir": task.base_dir,
                     "process": task.process,
-                    "script_path":task.script_path,
+                    "script_path": task.script_path,
                     "path": path
                 }
 
+                handler = task.get("handler")
+                if handler:
+                    cmd = Common.create_from_class_path(handler, [], kwargs)
+                else:
+                    # create a "custom" command that will act on the file
+                    cmd = CheckinCmd(**kwargs)
 
-                # create a "custom" command that will act on the file
-                cmd = CustomCmd(
-                        **kwargs
-                )
+                #print "Process [%s] checking in [%s]" % (os.getpid(), path)
                 cmd.execute()
 
                 # TEST
@@ -141,8 +158,9 @@ class WatchFolderFileActionThread(threading.Thread):
                 #    os.unlink(path)
 
                 count += 1
-                if count > 50:
+                if count == 20:
                     restart = True
+                    task.set_clean(True)
                     break
 
 
@@ -154,15 +172,32 @@ class WatchFolderFileActionThread(threading.Thread):
                 #raise
 
             finally:
-                os.unlink(checkin_path)
+
+                task.set_clean(True)
+                if os.path.exists(checkin_path):
+                    os.unlink(checkin_path)
+                if os.path.exists(lock_path):
+                    os.unlink(lock_path)
+                task.set_clean(False)
+                
+                if restart:
+                    task.set_clean(True)
+
+
 
         # restart every 20 check-ins
         if restart:
             for path in paths:
                 checkin_path = "%s.checkin" % path
+                lock_path = "%s.lock" % path
                 if os.path.exists(checkin_path):
                     os.unlink(checkin_path)
-            Common.restart()
+                if os.path.exists(lock_path):
+                    os.unlink(lock_path)
+            # this exaggerates the effect of not pausing check thread for cleaning
+            #time.sleep(10)
+            Common.kill()
+
 
 
 class WatchFolderCheckFileThread(threading.Thread):
@@ -181,26 +216,36 @@ class WatchFolderCheckFileThread(threading.Thread):
 
         try:
             path = my.kwargs.get("path")
-
-            if os.path.exists(my.lock_path):
+           
+            # this extra checkin_path check may not be needed
+            if os.path.exists(my.lock_path) or os.path.exists(my.checkin_path):
                 return
 
+            task = my.kwargs.get("task")
+ 
+            if task.in_clean():
+                return
+            
+            pid = os.getpid()
             f = open(my.lock_path, "w")
             f.close()
 
             changed = my.verify_file_size(path)
-
-
             if changed:
                 if os.path.exists(my.lock_path):
                     os.unlink(my.lock_path)
                 return
 
+            # time has passed, check again
+            if task.in_clean():
+                return
+
+
 
             f = open(my.checkin_path, "w")
+            f.write(str(pid))
             f.close()
 
-            task = my.kwargs.get("task")
             task.add_path(path)
 
 
@@ -238,7 +283,11 @@ class WatchFolderCheckFileThread(threading.Thread):
                 changed = True
                 break
 
-            file_size2 = os.path.getsize(file_path)
+            if os.path.isdir(file_path):
+                file_size2 = os.path.getsize(file_path)
+            else:
+                file_size2 = os.path.getsize(file_path)
+
             mtime2 = os.path.getmtime(file_path)
             #print "file_size2: ", file_size2
             #print "mtime2: ", mtime2
@@ -256,7 +305,8 @@ class WatchFolderCheckFileThread(threading.Thread):
 
 
 
-class CustomCmd(object):
+
+class CheckinCmd(object):
 
     def __init__(my, **kwargs):
         my.kwargs = kwargs
@@ -306,7 +356,7 @@ class CustomCmd(object):
         base_dir = my.kwargs.get("base_dir")
         search_type = my.kwargs.get("search_type")
         process = my.kwargs.get("process")
-        watch_script_path =  my.kwargs.get("script_path")
+        watch_script_path = my.kwargs.get("script_path")
         if not process:
             process = "publish"
 
@@ -334,7 +384,7 @@ class CustomCmd(object):
         server.set_project(project_code)
 
         transaction = Transaction.get(create=True)
-        server.start(title='Check-in of media', description='Check-in of media')
+        server.start(title='Check-in of %s'%search_type, description='Check-in of %s'%search_type)
 
         server_return_value = {}
 
@@ -371,14 +421,28 @@ class CustomCmd(object):
 
             #task = server.create_task(sobj.get('__search_key__'),process='publish')
             #server.update(task, {'status': 'New'})
-
-
+            
+            """
+            #TEST: simulate different check-in duration
+            from random import randint
+            sec = randint(1, 5)
+            print "checking in for ", sec, "sec"
+            server.eval("@SOBJECT(sthpw/login)")
+            import shutil
+            dir_name,base_name = os.path.split(file_path)
+            dest_dir = 'C:/ProgramData/Southpaw/watch_temp'
+            if not os.path.exists(dest_dir):
+                os.makedirs(dest_dir)
+            shutil.move(file_path, '%s/%s'%(dest_dir, base_name))
+            time.sleep(sec)
+            # move back the file in a few seconds 
+            shutil.move('%s/%s'%(dest_dir, base_name), file_path)
+            """
             server_return_value = server.simple_checkin(search_key,  context, file_path, description=description, mode='move')
 
             if watch_script_path:
                 cmd = PythonCmd(script_path=watch_script_path,search_type=search_type,drop_path=file_path,search_key=search_key)
                 cmd.execute()
-
 
 
 
@@ -431,10 +495,13 @@ class CustomCmd(object):
             f.write(pre_log)
             f.close()
 
-            # Delete the sourse file after check-in step.
-            print "File checked in."
+            # Delete the source file after check-in step.
+            print "File handled."
             if os.path.exists(file_path):
-                os.unlink(file_path)
+                if os.path.isdir(file_path):
+                    os.rmdirs(file_path)
+                else:
+                    os.unlink(file_path)
                 print "Source file [%s] deleted: " %file_name
 
 
@@ -447,16 +514,20 @@ class WatchDropFolderTask(SchedulerTask):
 
     def __init__(my, **kwargs):
 
+        my.input_kwargs = kwargs
         my.base_dir = kwargs.get("base_dir")
         my.project_code = kwargs.get("project_code")
         my.search_type = kwargs.get("search_type")
         my.process = kwargs.get("process")
         my.script_path = kwargs.get("script_path")
+        my.watch_folder_code = kwargs.get("watch_folder_code")
 
         super(WatchDropFolderTask, my).__init__()
 
         my.checkin_paths = []
+        my.in_clean_mode = False
 
+        my.files_checked = 0
 
     def add_path(my, path):
         my.checkin_paths.append(path)
@@ -465,7 +536,15 @@ class WatchDropFolderTask(SchedulerTask):
         return my.checkin_paths
 
 
+    def get(my, key):
+        return my.input_kwargs.get(key)
 
+
+    def set_clean(my, clean):
+        my.in_clean_mode = clean
+
+    def in_clean(my):
+        return my.in_clean_mode
 
     def _execute(my):
 
@@ -500,12 +579,13 @@ class WatchDropFolderTask(SchedulerTask):
         if not dirs:
             return
 
-        #print "Found new: ", dirs
+        
 
         # go thru the list to check each file
         for file_name in dirs:
-
             file_path = '%s/%s' %(my.base_dir, file_name)
+            if file_path in my.get_paths():
+                continue
             thread = WatchFolderCheckFileThread(
                     task=my,
                     path=file_path
@@ -522,12 +602,10 @@ class WatchDropFolderTask(SchedulerTask):
             print "WARNING: No base dir defined."
             return
 
-
-        # Start check-in thread
-        
+        # Start action thread
         checkin = WatchFolderFileActionThread(
                 task=my,
-                )
+        )
         checkin.start()
 
         # execute and react based on a loop every second
@@ -569,6 +647,8 @@ class WatchDropFolderTask(SchedulerTask):
 
         print "Running Watch Folder ..."
 
+
+
         # Check whether the user define the drop folder path.
         # Default dop folder path: /tmp/drop
         parser = OptionParser()
@@ -577,6 +657,11 @@ class WatchDropFolderTask(SchedulerTask):
         parser.add_option("-s", "--search_type", dest="search_type", help="Define search_type.")
         parser.add_option("-P", "--process", dest="process", help="Define process.")
         parser.add_option("-S", "--script_path",dest="script_path", help="Define script_path.")
+        parser.add_option("-w", "--watch_folder_code",dest="watch_folder_code", 
+				help="Define watch folder code. If no code is used, then it assumed that this process \
+				is managed in a standalone script.")
+        
+        parser.add_option("-c", "--handler",dest="handler", help="Define Custom Handler Class.")
         (options, args) = parser.parse_args()
 
         
@@ -586,7 +671,8 @@ class WatchDropFolderTask(SchedulerTask):
         if options.project != None :
             project_code= options.project
         else:
-            project_code= 'jobs'
+            raise Exception("No project specified")
+
 
         if options.drop_path!=None :
             drop_path= options.drop_path
@@ -600,23 +686,51 @@ class WatchDropFolderTask(SchedulerTask):
         if options.search_type!=None :
             search_type = options.search_type
         else:
-            search_type = 'jobs/media'
+            search_type = None
+
 
         if options.process!=None :
             process = options.process
         else:
             process= 'publish'
 
-        if options.script_path!=None :
+        if options.script_path != None :
             script_path = options.script_path
         else:
-            script_path="None"
+            script_path = None
           
+        
+
+        if options.handler != None:
+            handler = options.handler
+        else:
+            handler = None
+
+        if options.watch_folder_code != None:
+            watch_folder_code = options.watch_folder_code
+        else:
+            watch_folder_code = None
+        if watch_folder_code:   
+            # record pid in watch folder pid file
+            pid = os.getpid()
+            pid_file = "%s/log/watch_folder.%s" % (Environment.get_tmp_dir(), watch_folder_code)
+            f = open(pid_file, "w")
+            f.write(str(pid))
+            f.close()
+
+        Batch(project_code=project_code)
 
 
 
-
-        task = WatchDropFolderTask(base_dir=drop_path, project_code=project_code,search_type=search_type, process=process,script_path=script_path)
+        task = WatchDropFolderTask(
+            base_dir=drop_path, 
+            project_code=project_code,
+            search_type=search_type, 
+            process=process,
+            script_path=script_path, 
+            handler=handler,
+            watch_folder_code=watch_folder_code
+        )
         
         scheduler = Scheduler.get()
         scheduler.add_single_task(task, delay=1)
@@ -625,7 +739,6 @@ class WatchDropFolderTask(SchedulerTask):
     start = classmethod(start)
 
 if __name__ == '__main__':
-    Batch()
     WatchDropFolderTask.start()
     while 1:
         try:
